requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docstring-parser==0.15
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.8
Jinja2==3.1.2
jsonargparse==4.9.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lightning-flash==0.8.1.post0
lightning-utilities==0.8.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==2.3.0
torch==2.0.0
torchmetrics==0.10.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
Werkzeug==2.3.3
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
